feat: added evaluate_rpn function and filling in the skeleton
Need to add functionality for negative numbers in token strings as '4-' which actually means -4. Need to create parse_int(), int_to_string(), and potentially insert_queue() that allows a number to be added at a particular location in the queue. This will be important with the '-(' portion of a potential string